Immediately hide task/chore when tracking
This causes a chore or task to immediately be hidden once tracked,
rather than requiring the round-trip to grocy & back to homeassistant
to complete. It only hides the task/chore until the next refresh from
grocy, at which point that task/chore will be updated appropriately (and
therefore won't be shown in the visible state).