Fix for Issue #163 and #164
I tried to fix the issues i mentioned myself.
Code for Custom sort is modified with the help of ChatGPT.
And the code to add task category while creating a new task is put together from various other parts of the existing code.

Might need a bit of cleanup.
Kindly review this.
Everything seems to be working in my testing.
The only thing i currently have doubt is:
 - When there is no task category in grocy. (The select dropdown for category will be blank. Hopefully that doesn't cause any errors). If someone can test that please.